[GWFlow] moved to subdirectory, put into separate namespace
